requirements.txt
pytesseract
pillow
pandas
tabula-py
openpyxl
pymupdf
